requirements.txt
Django==2.1.7  
djangorestframework==3.9.2  
numpy==1.16.2 
pandas==0.24.1 
pip==19.0.3 
psycopg2==2.7.7  
psycopg2-binary==2.7.7  
Pygments==2.3.1  
python-dateutil==2.8.0  
pytz==2018.9 
scikit-learn==0.20.3 
scipy==1.2.1  
setuptools==40.8.0 
six==1.12.0 
wheel==0.33.1 
